+++ homework/console/03/@types/chartjs/index.d.ts
@@ -1,0 +1,1 @@
+declare module 'chartjs';
