+++ homework/console/03/index.html
@@ -48,11 +48,7 @@
       </div>
     </div>
   </main>
-  <script src="https://unpkg.com/axios/dist/axios.min.js"></script>
-  <script src="https://cdn.jsdelivr.net/npm/chart.js@2.8.0"></script>
-  <script>var exports = {};</script>
-  <script src="./app.js" type="module"></script>
-  <script src="./app.ts" type="module"></script>
+  <script src="./src/app.ts" type="module"></script>
 
 </body>
