+++ package.json
@@ -35,5 +35,9 @@
     "prettier-eslint": "^13.0.0",
     "ts-node": "^10.7.0",
     "typescript": "^4.6.2"
+  },
+  "dependencies": {
+    "axios": "^0.26.1",
+    "chartjs": "^0.3.24"
   }
 }
